requirements.txt
# AI和机器学习相关 - 暂时注释
# tensorflow-cpu==2.5.0
# scikit-learn==0.24.2
# pandas==1.3.3
# numpy==1.21.2

# Web框架和API
fastapi==0.68.0
uvicorn==0.15.0
pydantic==1.8.2
python-multipart==0.0.5

# 地理信息处理 - 暂时注释
# geopandas==0.13.2
# folium==0.14.0
# osmnx==1.6.0
# rtree==1.0.1  # 空间索引支持

# 数据库
sqlalchemy==1.4.23
# pymongo==4.5.0

# 数据可视化 - 暂时注释
# plotly==5.16.1
# dash==2.13.0

# 工具库
python-dotenv==0.19.0
requests==2.26.0
# pytest==7.4.2
# black==23.7.0
aiohttp==3.8.1
jinja2==3.0.1
python-jose==3.3.0
passlib==1.7.4
bcrypt==3.2.0
aiofiles==0.7.0
email-validator==1.1.3
polyline==1.4.0 